WebApp Release Candidate v1. Reimplemented Rest Client. Refactored all Web Entitles according to specs. Bug fixes and clean-up. Updated styling.

+++ NoNameWebApp/NoNameWebApp/Presentation/CommonPresentationStuff.cs
@@ -71,7 +71,7 @@ namespace NoNameWebApp.Presentation
             ShowAlert(page, "Greška u komunikaciji sa servisom!");
         }
 
-        public static string FormatPrice(float price)
+        public static string FormatPrice(double price)
         {
             return string.Format("{0:0.00} kn", price);
         }
